requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# CrewAI Integration
crewai[tools]>=0.117.0
crewai-tools>=0.12.0

# AI & ML
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data Sources
vnstock>=3.2.0
yfinance>=0.2.0
requests>=2.31.0
aiohttp>=3.8.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional (for advanced features)
# asyncio-mqtt>=0.11.0  # Uncomment if using MQTT
# redis>=4.5.0          # Uncomment if using Redis cache
# celery>=5.3.0         # Uncomment if using background tasks